refactor: Add plugins/indext.ts and plugins/registry.ts
